backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# API Proxy
requests>=2.30.0,<3.0

# Production / Deployment
gunicorn>=21.0.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
whitenoise>=6.6.0
python-dotenv>=1.0.1
